TALON BOT MARK 2 github/requirements.txt
discord.py==2.3.2
virtualenv==20.24.2